backend/pyproject.toml
[project]
name = "sheleads-backend"
version = "0.1.0"
description = "SheLeads PPT Backend - FastAPI with Clerk Auth, Ollama RAG"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "motor>=3.3.0",
    "python-dotenv>=1.0.0",
    "ollama>=0.3.0",
    "qdrant-client",
    "pydantic>=2.8.0",
    "httpx>=0.27.0",
    "PyJWT>=2.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
